contract_audit/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-docx==1.1.0
pymupdf==1.23.26
httpx==0.26.0
aiosqlite==0.19.0
pydantic==2.6.1
python-multipart==0.0.9
